backend/requirements.txt
flask==2.1.0
flask-cors==3.0.10
werkzeug==2.1.1
python-docx==0.8.11
python-magic-bin==0.4.24
openai>=0.27.8
python-dotenv==0.20.0
PyPDF2==2.11.1